# doc/api.yaml
openapi: 3.0.3
info:
  title: Simplified login API specification
  description: |-
    This OpenAPI document describes the simplified login API.
    Copy and paste the API from the `paths` key to your OpenAPI document.
  version: "1"
components:
  securitySchemes:
    bearerAuth: # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT # optional, arbitrary value for documentation purposes
  responses:
    InvalidRequestError:
      description: bad request
      content:
        application/json:
          schema:
            description: '400 error'
            type: object
            properties:
              error_message:
                description: text of error message
                type: string
                pattern: '^.*?$'
                minLength: 1
                maxLength: 30
                example: "Bad request, retry!" 
    ForbiddenError:
      description: Access token is missing or invalid
      content:
        application/json:
          schema:
            description: '403 error'
            type: object
            properties:
              error_message:
                description: text of error message
                type: string
                pattern: '^.*?$'
                minLength: 1
                maxLength: 50
                example: "Access token is missing or invalid!" 
    InternalServerError:
      description: server error
      content:
        application/json:
          schema:
            description: '500 error'
            type: object
            properties:
              error_message:
                description: text of error message
                type: string
                pattern: '^.*?$'
                minLength: 1
                maxLength: 30
                example: "Server error, try again later!"
    NotFoundError:
      description: object not found, due to a wrong request
      content:
        application/json:
          schema:
            description: '404 error'
            type: object
            properties:
              error_message:
                description: text of error message
                type: string
                pattern: '^.*?$'
                minLength: 1
                maxLength: 30
                example: "The object is not found!"
  schemas:
    GeneralId:
      type: object
      description: id of a user, a conversation, a group, a message or a comment
      properties:
        value:
          type: integer
          description: id value
          minimum: 0
          maximum: 10000
    Username:
      type: object
      description: username
      properties:
        name:
          type: string
          description: | 
            text that represents the username 
          pattern: '^[a-z0-9]{1,15}$'
          minLength: 1
          maxLength: 15
      required: [name]
    Message:
      type: object
      description: message of a conversation
      properties:
        id:
          $ref: '#/components/schemas/GeneralId'
        date_time:
          type: string
          description: date and time of the message
          format: date-time
          pattern: '^\\d{4}-(0[1-9]|1[0-2])-(0[1-9]|[12]\\d|3[01])T([01]\\d|2[0-3]):[0-5]\\d$'
          minLength: 16
          maxLength: 16
        text:
          description: text of the message
          type: string
          pattern: '^.*?$'
          minLength: 1
          maxLength: 9999
        status:
          type: boolean
          description: status of the message
        conv_id:
          $ref: '#/components/schemas/GeneralId'
        sender_id:
          $ref: '#/components/schemas/GeneralId'
        photo:
          type: string
          description: photo of the message
          format: byte
          pattern: '^.*?$'
          minLength: 1
          maxLength: 7000000
        gif:
          type: string
          description: gif of the message
          format: byte
          pattern: '^.*?$'
          minLength: 1
          maxLength: 7000000
        forward:
          type: boolean
          description: define if is a forwarded message
        reply_id:
          $ref: '#/components/schemas/GeneralId'
...
    User:
      type: object
      description: general user
      properties:
        id:
          $ref: '#/components/schemas/GeneralId'
        user_name:
          $ref: '#/components/schemas/Username'
        photo:
          type: string
          description: group photo
          pattern: '^.*?$'
          format: byte
          minLength: 1
          maxLength: 7000000
      required: [id, user_name]
...
paths:
...
  /users:
    description: area where is possibile to search users
    get:
      security:
        - bearerAuth: []
      tags: ["search users"]
      operationId: searchUsers
      summary: get a list of users
      description: get a list of users who have the username starting with an input string
      responses:
        '200':
          description: |
            an array of object, each one contains a name, a message,
            a array users and a photo
          content:
            application/json:
              schema:
                type: object
                description: '200 response message'
                properties: 
                  filtered_users:
                    description: conversations array in output
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
                    minItems: 0
                    maxItems: 10000
        '400':
          $ref: '#/components/responses/InvalidRequestError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /users/{user}/conversations/{conv_id}/messages:
    description: messages area, where user can send text, photo and gif
    parameters:
      - name: user
        description: user id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/GeneralId'
      - name: conv_id
        description: conv id 
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/GeneralId'
    post:
      security:
        - bearerAuth: []
      tags: ["sending a message"]
      operationId: sendMessage
      summary: sends a message (so, adds it to list of messages a conversation)
      description: | 
        User sends a message to the other user/s
      requestBody:
        content:
          multipart/form-data::
            schema:
              type: object
              description: object that contains message input data
              properties:
                text:
                  type: string
                  description: text of the message
                  pattern: '^.*?$'
                  minLength: 1
                  maxLength: 9999
                messToReplyTo:
                  $ref: '#/components/schemas/GeneralId'
                image:
                  description: new image in input
                  type: string
                  pattern: '^.*?$'
                  format: byte
                  minLength: 1
                  maxLength: 7000000
                gif:
                  description: new gif in input
                  type: string
                  pattern: '^.*?$'
                  format: byte
                  minLength: 1
                  maxLength: 7000000
        required: true
      responses:
        '201':
          description: New message sending action successful
          content:
            application/json:
              schema:
                type: object
                description: object that contains message input data
                properties:
                  mess:
                    $ref: '#/components/schemas/Message'
                  mess_to_reply:
                    $ref: '#/components/schemas/Message'
        '400':
          $ref: '#/components/responses/InvalidRequestError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /users/{user}/conversations/{conv_id}/messages/{mess_id}:
    description: message area, where user can forward or delete it
    parameters:
      - name: user
        description: user id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/GeneralId'
      - name: conv_id
        description: receiver id (group or user)
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/GeneralId'
      - name: mess_id
        description: message id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/GeneralId'
    post:
      security:
        - bearerAuth: []
      tags: ["forwarding of a message"]
      operationId: forwardMessage
      summary: forwards the message
      description: User forwards a message in a chat, represented by his id
      requestBody:
        description: User gives the chat id in input
        content:
          application/json:
            schema:
              type: object
              description: object that contains array of conversations data
              properties:
                destination:
                  description: conversations array in input
                  type: array
                  items:
                    type: object
                    description: object with all data of a conversation
                    properties: 
                      user_id:
                        $ref: '#/components/schemas/GeneralId'
                      group_id:
                        $ref: '#/components/schemas/GeneralId'
                  minItems: 0
                  maxItems: 10000
      responses:
        '201':
          description: message forwarding action successful
          content:
            application/json:
              schema:
                description: '201 response message'
                type: object
                properties:
                  messages:
                    description: '201 response message'
                    type: array
                    items:
                      $ref: '#/components/schemas/Message'
                    minItems: 0
                    maxItems: 10000
        '400':
          $ref: '#/components/responses/InvalidRequestError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      security:
        - bearerAuth: []
      tags: ["deleting of a message"]
      operationId: deleteMessage
      summary: deletes the message
      description: User deletes a message
      responses:
        '200':
          description: message deletion action successful
          content:
            plain/text:
              schema:
                description: '200 response message'
                type: string
                pattern: '^.*?$'
                minLength: 1
                maxLength: 30
                example: "Deletion accepted!"
        '400':
          $ref: '#/components/responses/InvalidRequestError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '500':
          $ref: '#/components/responses/InternalServerError'
...